+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Models/ApproverType.cs
@@ -1,0 +1,41 @@
+//    #[license]
+//    SmartsheetClient SDK for C#
+//    %%
+//    Copyright (C) 2014 SmartsheetClient
+//    %%
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//        
+//            http://www.apache.org/licenses/LICENSE-2.0
+//        
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+//    %[license]
+
+namespace Smartsheet.Api.Models
+{
+    /// <summary>
+    /// Represents the type of approver who can approve a classification downgrade.
+    /// </summary>
+    public enum ApproverType
+    {
+        /// <summary>
+        /// The approvers are groups.
+        /// </summary>
+        GROUPS,
+
+        /// <summary>
+        /// The approvers are users.
+        /// </summary>
+        USERS,
+
+        /// <summary>
+        /// The approvers are the workspace admins.
+        /// </summary>
+        WORKSPACE_ADMINS
+    }
+}